requirements.txt
python-dotenv==1.0.0
sqlalchemy==2.0.23
pyodbc==4.0.39
pandas==2.1.1 